requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==1.4.53
pydantic==1.10.16
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
pytest-cov==4.1.0
psycopg2-binary==2.9.10
bcrypt==4.0.1
email-validator>=1.1.1,<2.0.0